+++ avatars/api.py
@@ -1,5 +1,5 @@
 # This file has been generated - DO NOT MODIFY
-# API Version : 5.1.0-5d5865f0a4b03a10f73a0894938c6843af952096
+# API Version : 8.0.0
 
 
 import logging