+++ avatars/arrow_utils.py
@@ -17,18 +17,24 @@ from typing import (
     Union,
 )
 
+import magic
 import pandas as pd
 import pyarrow as pa
 import pyarrow.csv as pcsv
 import pyarrow.dataset as ds
 import pyarrow.ipc as pi
 import pyarrow.parquet as pq
+import structlog
 
+ERROR_PRE = "Invalid Arrow stream - "
 PARQUET_MARKER = b"PAR1"
-ARROW_STREAM_MARKER = b"ARRS1"
+ARROW_BEGIN_OF_STREAM_MARKER = b"ARRBOS1"
+ARROW_END_OF_STREAM_MARKER = b"ARREOS1"
 ARROW_BATCH_MARKER = b"ARRB1"
 DEFAULT_MAX_ROWS_PER_BATCH = 1_000_000
 APPLICATION_ARROW_STREAM = "application/vnd.apache.arrow.stream"
+
+DEFAULT_MAGIC_BYTES = 2048
 
 
 BatchBytesFunction = Callable[[bytes], None]
@@ -41,9 +47,12 @@ TableSource = Union[str, FileLike]
 DataSourceItem = Union[TableSource, pa.Table]
 DataSourceItems = Union[DataSourceItem, list[DataSourceItem]]
 
+logger = structlog.get_logger(__name__)
+
 
 class StreamState(Enum):
-    WAIT_STREAM = "stream"
+    WAIT_STREAM_BEGIN = "stream-begin"
+    AT_STREAM_END = "stream-end"
     WAIT_BATCH = "batch"
     DATA = "data"
 
@@ -80,6 +89,24 @@ def is_text_file_or_buffer(obj: Any) -> bool:
     return is_text_file(obj) or is_text_buffer(obj)
 
 
+def guess_file_format(obj: Any) -> str:
+    fmt: Optional[str] = None
+
+    if is_file_like(obj):
+        fmagic = magic.from_buffer(obj.read(DEFAULT_MAGIC_BYTES))
+        obj.seek(0, os.SEEK_SET)
+    else:
+        fmagic = magic.from_file(obj)
+
+    if fmagic:
+        if "Parquet" in fmagic:
+            fmt = "parquet"
+        elif "CSV" in fmagic or "ASCII text" in fmagic:
+            fmt = "csv"
+
+    return fmt or fmagic or "unknown"
+
+
 def flatten(items: Any) -> Iterator[Any]:
     if isinstance(items, list):
         for item in items:
@@ -114,7 +141,7 @@ def available(data: BytesIO) -> int:
         return data.tell()
 
 
-class HeaderBase:
+class MarkerBase:
     def __init__(self, *, total_size: int) -> None:
         self.total_size = total_size
 
@@ -126,9 +153,9 @@ class HeaderBase:
 
     def check_can_load(self, data: memoryview) -> None:
         if not self.can_load(data):
-            raise RuntimeError(
+            raise ValueError(
                 f"Expected at least {self.total_size} bytes, "
-                f" got {len(data)} insrtead"
+                f"got {len(data)} instead"
             )
 
     def load(self, data: memoryview) -> memoryview:
@@ -147,7 +174,7 @@ class HeaderBase:
         return data[size:]
 
 
-class Header(HeaderBase):
+class Marker(MarkerBase):
     def __init__(self, marker: bytes, *, aux_size: int = 0) -> None:
         self.marker = marker
         self.marker_size = len(self.marker)
@@ -157,6 +184,12 @@ class Header(HeaderBase):
     def make(self) -> bytes:
         return super().make() + self.marker
 
+    def identify(self, data: memoryview) -> bool:
+        if self.can_load(data):
+            return self.check_marker(data, expected=False)
+
+        return False
+
     def load(self, data: memoryview) -> memoryview:
         # Load "my" parts (marker) and return the rest
         my_data = super().load(data)
@@ -164,46 +197,60 @@ class Header(HeaderBase):
 
         return self.skip_data(my_data, self.marker_size)
 
-    def check_marker(self, data: memoryview) -> None:
+    def check_marker(self, data: memoryview, *, expected: bool = True) -> bool:
         marker = data[0 : self.marker_size]
 
         if marker != self.marker:
-            raise ValueError(
-                f"Expected '{str(self.marker)}', got {str(marker)} instead"
-            )
+            if expected:
+                raise ValueError(
+                    f"{ERROR_PRE}"
+                    f"Expected '{self.marker!r}', got {bytes(marker)!r} instead"
+                )
+            else:
+                return False
+
+        return True
 
 
-class HeaderWithSize(Header):
+class MarkerWithSize(Marker):
     size_format = "!I"
     size_packed_size = struct.calcsize(size_format)
 
     def __init__(self, marker: bytes, size: int = 0) -> None:
         self.size = size
-        super().__init__(marker, aux_size=HeaderWithSize.size_packed_size)
+        super().__init__(marker, aux_size=MarkerWithSize.size_packed_size)
 
     def make(self) -> bytes:
         return super().make() + struct.pack(self.size_format, self.size)
 
     def load(self, data: memoryview) -> memoryview:
         my_data = super().load(data)
-        packed_size = my_data[0 : HeaderWithSize.size_packed_size]
+        packed_size = my_data[0 : MarkerWithSize.size_packed_size]
 
-        self.size = struct.unpack(HeaderWithSize.size_format, packed_size)[0]
+        self.size = struct.unpack(MarkerWithSize.size_format, packed_size)[0]
 
-        return self.skip_data(my_data, HeaderWithSize.size_packed_size)
+        return self.skip_data(my_data, MarkerWithSize.size_packed_size)
 
 
 class ArrowDatasetBuilder:
     def __init__(self) -> None:
         self.inferred_type: Optional[str] = None
 
-    def to_table(self, source: TableSource) -> pa.Table:
-        try:
-            self.inferred_type = "csv"
+    def read_table(self, source: TableSource, fmt: str) -> pa.Table:
+        if fmt == "csv":
             return pcsv.read_csv(source)
-        except:  # noqa: E722
-            self.inferred_type = "parquet"
+        elif fmt == "parquet":
             return pq.read_table(source)
+
+        raise ValueError(f"{ERROR_PRE}" f"Expected known format, got '{fmt}' instead")
+
+    def to_table(self, source: TableSource) -> pa.Table:
+        fmt = guess_file_format(source)
+        self.inferred_type = fmt
+
+        logger.info(f"dataset item guessed format is {fmt}")
+
+        return self.read_table(source, fmt)
 
     def to_source(self, item: Any) -> Union[str, pa.Table]:
         if p := is_existing_file_or_dir(item):
@@ -243,10 +290,11 @@ class ArrowStreamReader:
         self.func_stack: list[Any] = []
         self.batch = bytes()
         self.data = bytearray()
-        self.stream_header = Header(ARROW_STREAM_MARKER)
-        self.batch_header = HeaderWithSize(ARROW_BATCH_MARKER)
-        self.state = StreamState.WAIT_STREAM
-        self.offset = 0
+        self.stream_header = Marker(ARROW_BEGIN_OF_STREAM_MARKER)
+        self.stream_footer = MarkerWithSize(ARROW_END_OF_STREAM_MARKER)
+        self.batch_header = MarkerWithSize(ARROW_BATCH_MARKER)
+        self.state = StreamState.WAIT_STREAM_BEGIN
+        self.nb_batches = 0
 
     def __enter__(self) -> Any:
         return self
@@ -316,20 +364,32 @@ class ArrowStreamReader:
         self.table_func = table_func
 
     def process_end(self) -> None:
+        self.update_state()
+        self.check_stream_started()
         self.flush()
+        self.check_at_stream_end()
         self.check_no_remaining_data()
+        self.check_nb_batches()
         self.pop_funcs()
 
     def process_chunk(self, chunk: bytes) -> None:
-        self.data.extend(chunk)
+        if len(chunk) == 0:
+            return
 
+        if self.state == StreamState.AT_STREAM_END:
+            raise ValueError("Receiving data after stream end")
+
+        self.data.extend(chunk)
+        self.update_state()
+
+    def update_state(self) -> None:
         while True:
             prev_state = self.state
 
-            if self.state == StreamState.WAIT_STREAM:
-                self.load_header(self.stream_header, StreamState.WAIT_BATCH)
+            if self.state == StreamState.WAIT_STREAM_BEGIN:
+                self.try_load_header(self.stream_header, StreamState.WAIT_BATCH)
             elif self.state == StreamState.WAIT_BATCH:
-                self.load_header(self.batch_header, StreamState.DATA)
+                self.try_load_header(self.batch_header, StreamState.DATA)
             elif self.state == StreamState.DATA:
                 self.flush()
 
@@ -337,10 +397,23 @@ class ArrowStreamReader:
                 # No progress in this loop, need more data
                 break
 
-    def load_header(self, header: HeaderBase, new_state: StreamState) -> None:
-        if header.try_load(memoryview(self.data)):
-            self.skip_data(header.total_size)
-            self.state = new_state
+    def try_load_header(self, header: Marker, new_state: StreamState) -> None:
+        data = memoryview(self.data)
+
+        if self.stream_footer.identify(data):
+            self.stream_footer.load(data)
+            # End of stream
+            self.skip_marker_and_set_state(
+                self.stream_footer, StreamState.AT_STREAM_END
+            )
+        elif header.try_load(data):
+            self.skip_marker_and_set_state(header, new_state)
+
+    def skip_marker_and_set_state(
+        self, marker: MarkerBase, new_state: StreamState
+    ) -> None:
+        self.skip_data(marker.total_size)
+        self.state = new_state
 
     def flush(self) -> None:
         if not self.is_batch_complete() or self.no_data():
@@ -348,6 +421,7 @@ class ArrowStreamReader:
 
         # Read batch bytes
         self.batch = self.extract_data(self.batch_header.size)
+        self.nb_batches += 1
 
         # Keep remaining data from next batch
         self.skip_data(self.batch_header.size)
@@ -390,7 +464,24 @@ class ArrowStreamReader:
         remaining = len(self.data)
 
         if remaining:
-            raise RuntimeError(f"Expected no remaining data, got {remaining} bytes")
+            raise ValueError(f"Expected no remaining data, got {remaining} bytes")
+
+    def check_nb_batches(self) -> None:
+        if self.nb_batches != self.stream_footer.size:
+            raise ValueError(
+                f"Expected {self.stream_footer.size} batches, "
+                f"got {self.nb_batches} instead"
+            )
+
+    def check_stream_started(self) -> None:
+        if self.state == StreamState.WAIT_STREAM_BEGIN:
+            raise ValueError(f"Expected a stream, got {bytes(self.data)!r} instead")
+
+    def check_at_stream_end(self) -> None:
+        if self.state != StreamState.AT_STREAM_END:
+            raise ValueError(
+                "Expected to be at stream end, " f"got {self.state.value} instead"
+            )
 
 
 class ArrowStreamWriter:
@@ -399,8 +490,9 @@ class ArrowStreamWriter:
     ) -> None:
         self.ds = ds
         self.nb_rows_per_batch = nb_rows_per_batch
-        self.batch_header = HeaderWithSize(ARROW_BATCH_MARKER)
+        self.batch_header = MarkerWithSize(ARROW_BATCH_MARKER)
         self.batch: Optional[bytes] = None
+        self.nb_batches = 0
         self.reset_writer()
 
     def __enter__(self) -> Any:
@@ -409,13 +501,22 @@ class ArrowStreamWriter:
     def __exit__(self, *args: Any) -> Any:
         pass
 
-    def __iter__(self) -> Iterator[bytes]:
-        yield Header(ARROW_STREAM_MARKER).make()
+    def iter_core(self) -> Iterator[bytes]:
+        yield Marker(ARROW_BEGIN_OF_STREAM_MARKER).make()
 
-        for batch in self.ds.to_batches():
+        for batch in self.ds.to_batches(batch_size=self.nb_rows_per_batch):
             yield from self.process_batch(batch)
 
         yield from self.flush()
+
+        yield MarkerWithSize(ARROW_END_OF_STREAM_MARKER, self.nb_batches).make()
+
+    async def __aiter__(self) -> Stream:
+        for v in self.iter_core():
+            yield v
+
+    def __iter__(self) -> Iterator[bytes]:
+        yield from self.iter_core()
 
     def process_batch(self, batch: pa.RecordBatch) -> Iterator[bytes]:
         if self.should_flush(batch):
@@ -430,6 +531,8 @@ class ArrowStreamWriter:
     def flush(self) -> Iterator[bytes]:
         self.close_writer()
         self.reset_writer()
+        self.nb_batches += 1
+
         yield from self.get_buffers()
 
     def get_buffers(self) -> Iterator[bytes]:
@@ -443,7 +546,8 @@ class ArrowStreamWriter:
 
     def close_writer(self) -> None:
         self.writer.close()
-        self.batch = self.buf.getvalue()
+        buffer = self.buf.getvalue()
+        self.batch = buffer if len(buffer) else None
 
     def reset_writer(self) -> None:
         self.buf = io.BytesIO()