+++ notebooks/Tutorial10_Geolocation_privacy_assessment.py
@@ -5,7 +5,7 @@
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
-#       jupytext_version: 1.16.1
+#       jupytext_version: 1.16.2
 # ---
 
 # # Tutorial 10: Geolocation privacy assessment